recent songs
The program will now get a csv file of the 50 most recently played songs, and it determines whether the song is in your "liked songs". This can now be used for machine learning, given we can try to predict if I might like a song based on the four features.